requirements.txt
Flask>=3.0.0
python-dotenv>=1.0.0
flask-cors>=4.0.0
pandas>=2.1.0
scikit-learn>=1.3.0
numpy>=1.26.0
gunicorn>=22.0.0
PyYAML>=6.0
